Move function that adapts the tokenizer
