--- outlines/integrations/__init__.py
@@ -1,1 +1,0 @@
-"""Utility functions and classes used to integrate `outlines` into other packages."""
