--- outlines/serve/vllm.py
@@ -1,4 +1,0 @@
-from outlines.integrations.vllm import (  # noqa[F401]
-    JSONLogitsProcessor,
-    RegexLogitsProcessor,
-)
